Resolve ESLint max-classes-per-file issues
